requirements.txt
nltk
gensim==3.5.0
matplotlib==2.2.2
msgpack-numpy==0.4.1
msgpack-python==0.5.6
numpy==1.15.2
Pyro4==4.72
scipy==1.1.0
stop-words==2015.2.23.1
pyYAML==3.13

django==2.1.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.3
django-allauth==0.38.0
django-cors-headers==2.4.0
djongo==1.2.31
